Added Shrine/Altar removal, reorganized some things, small bug fixes.

+++ core/src/com/censoredsoftware/Demigods/Engine/Tracked/TrackedBlock.java
@@ -82,7 +82,7 @@ public class TrackedBlock
 
 	public void remove()
 	{
-		getLocation().getBlock().setTypeIdAndData((byte) this.previousMaterial, (byte) this.previousMaterialByte, true);
+		getLocation().getBlock().setTypeIdAndData(this.previousMaterial, (byte) this.previousMaterialByte, true);
 		DemigodsData.jOhm.delete(TrackedBlock.class, getId());
 	}
